# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Testdrum",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/task/sort": {
            "post": {
                "tags": [
                    "Task"
                ],
                "summary": "Sort task",
                "description": "Sort tasks",
                "operationId": "0f4995ea5bfbe2b0549f2af96a684840",
                "requestBody": {
                    "description": "Pass params",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "status": {
                                        "type": "string"
                                    },
                                    "priority": {
                                        "type": "integer"
                                    },
                                    "title": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "copmleted_at": {
                                        "type": "string",
                                        "format": "date-time"
                                    },
                                    "created_at": {
                                        "type": "string",
                                        "format": "date-time"
                                    },
                                    "sortByCreatedAt": {
                                        "type": "string"
                                    },
                                    "sortByCompletedAt": {
                                        "type": "string"
                                    },
                                    "sortByPriority": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "tasks": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "status": {
                                                        "type": "string"
                                                    },
                                                    "priority": {
                                                        "type": "integer"
                                                    },
                                                    "title": {
                                                        "type": "string"
                                                    },
                                                    "description": {
                                                        "type": "string"
                                                    },
                                                    "copmleted_at": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    },
                                                    "parent_id": {
                                                        "type": "integer"
                                                    },
                                                    "depth": {
                                                        "type": "integer"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "You are not authorized to access this page."
                    }
                }
            }
        },
        "/api/task/{id}": {
            "get": {
                "tags": [
                    "Task"
                ],
                "summary": "task",
                "description": "Get task",
                "operationId": "90e79f793edef133aab13874da282d0f",
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "tasks": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "status": {
                                                        "type": "string"
                                                    },
                                                    "priority": {
                                                        "type": "integer"
                                                    },
                                                    "title": {
                                                        "type": "string"
                                                    },
                                                    "description": {
                                                        "type": "string"
                                                    },
                                                    "copmleted_at": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    },
                                                    "parent_id": {
                                                        "type": "integer"
                                                    },
                                                    "depth": {
                                                        "type": "integer"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Task"
                ],
                "summary": "Delete task",
                "description": "Delete task",
                "operationId": "5a7556048c897ff610dc0b9c0ac19488",
                "parameters": [
                    {
                        "name": "TaskId",
                        "in": "path",
                        "description": "TaskId",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "1, 2"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success"
                    }
                }
            }
        },
        "/api/task": {
            "put": {
                "tags": [
                    "Task"
                ],
                "summary": "Update task",
                "description": "Update task",
                "operationId": "db8432f5675244658667c51595ff3285",
                "requestBody": {
                    "description": "Pass params",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "priority": {
                                        "type": "integer"
                                    },
                                    "title": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "task_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "tasks": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "status": {
                                                        "type": "string"
                                                    },
                                                    "priority": {
                                                        "type": "integer"
                                                    },
                                                    "title": {
                                                        "type": "string"
                                                    },
                                                    "description": {
                                                        "type": "string"
                                                    },
                                                    "copmleted_at": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    },
                                                    "parent_id": {
                                                        "type": "integer"
                                                    },
                                                    "depth": {
                                                        "type": "integer"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "This task already done"
                    }
                }
            },
            "post": {
                "tags": [
                    "Task"
                ],
                "summary": "Sort task",
                "description": "Sort tasks",
                "operationId": "0ff66d3da3b3930ea4e952682522438c",
                "requestBody": {
                    "description": "Pass params",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "priority": {
                                        "type": "integer"
                                    },
                                    "title": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "task_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "tasks": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "status": {
                                                        "type": "string"
                                                    },
                                                    "priority": {
                                                        "type": "integer"
                                                    },
                                                    "title": {
                                                        "type": "string"
                                                    },
                                                    "description": {
                                                        "type": "string"
                                                    },
                                                    "copmleted_at": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    },
                                                    "parent_id": {
                                                        "type": "integer"
                                                    },
                                                    "depth": {
                                                        "type": "integer"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/task/{id}/complete": {
            "post": {
                "tags": [
                    "Task"
                ],
                "summary": "Complete task",
                "description": "Complete task",
                "operationId": "af2bf9d492e1eb701e26a9cb35b19c46",
                "requestBody": {
                    "description": "Pass params",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "priority": {
                                        "type": "integer"
                                    },
                                    "title": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "task_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "tasks": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "status": {
                                                        "type": "string"
                                                    },
                                                    "priority": {
                                                        "type": "integer"
                                                    },
                                                    "title": {
                                                        "type": "string"
                                                    },
                                                    "description": {
                                                        "type": "string"
                                                    },
                                                    "copmleted_at": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    },
                                                    "parent_id": {
                                                        "type": "integer"
                                                    },
                                                    "depth": {
                                                        "type": "integer"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Child task not done"
                    }
                }
            }
        }
    }
}